Leads/pyproject.toml
[tool.poetry]
name = "leads"
version = "0.1.0"
description = ""
authors = ["benie254 <davinci.monalissa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
dj-database-url = "^1.0.0"
django-bootstrap-v5 = "^1.0.11"
django-cors-headers = "^3.13.0"
django-rest-knox = "^4.2.0"
djangorestframework = "^3.14.0"
sendgrid = "^6.9.7"
psycopg2-binary = "^2.9.5"
whitenoise = {extras = ["brotli"], version = "^6.2.0"}
gunicorn = "^20.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


